requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
jinja2==3.1.6
python-multipart==0.0.20
openpyxl==3.1.5
pydantic==2.11.7
itsdangerous==2.2.0
qrcode[pil]==8.2
python-dotenv==1.1.1
psycopg[binary]==3.2.9
httpx==0.28.1
cryptography==42.0.5
